changed file name, add signout function
